* created a buzz.js file which handles buzzing and vibration reminders * cleaned up app/index.js
